+++ codeflash/verification/equivalence.py
@@ -3,7 +3,7 @@ import sys
 
 from codeflash.cli_cmds.console import console, logger
 from codeflash.verification.comparator import comparator
-from codeflash.verification.test_results import TestResults, TestType, VerificationType
+from codeflash.models.models import TestResults, TestType, VerificationType
 
 INCREASED_RECURSION_LIMIT = 5000
