+++ codeflash/verification/parse_test_output.py
@@ -20,8 +20,9 @@ from codeflash.code_utils.code_utils import (
     module_name_from_file_path,
 )
 from codeflash.discovery.discover_unit_tests import discover_parameters_unittest
-from codeflash.models.models import CoverageData, TestFiles
-from codeflash.verification.test_results import (
+from codeflash.models.models import (
+    CoverageData,
+    TestFiles,
     FunctionTestInvocation,
     InvocationId,
     TestResults,
