+++ tests/test_comparator.py
@@ -14,7 +14,7 @@ import pytest
 from codeflash.either import Failure, Success
 from codeflash.verification.comparator import comparator
 from codeflash.verification.equivalence import compare_test_results
-from codeflash.verification.test_results import FunctionTestInvocation, InvocationId, TestResults, TestType
+from codeflash.models.models import FunctionTestInvocation, InvocationId, TestResults, TestType
 
 
 def test_basic_python_objects() -> None:
