+++ tests/test_critic.py
@@ -11,7 +11,7 @@ from codeflash.models.models import (
     OptimizedCandidateResult,
 )
 from codeflash.result.critic import coverage_critic, performance_gain, quantity_of_tests_critic, speedup_critic
-from codeflash.verification.test_results import FunctionTestInvocation, InvocationId, TestResults, TestType
+from codeflash.models.models import FunctionTestInvocation, InvocationId, TestResults, TestType
 
 
 def test_performance_gain() -> None:
